Update patch set 1
Patch Set 1:

(8 comments)

Unlike .hgignore, .gitignore doesn't seem to be using regular expression. Please see:
https://git-scm.com/docs/gitignore

I added some suggestions in the comments, but I have not tried it myself so all comments start with "probably".

I think we don't really need to fully test everything. The file is just a way to automatically ignore build artifacts to prevent accidental git check in. This is going to be continuously revised. So long as we use the correct format (i.e. without regular expression) we can check in this patch.

Patch-set: 1
Label: Verified=0