Drop legacy code
C and perl code in this repository is not actively
maintained/built/used. It is not part of CI tests and
the build scripts are not tested on modern version of supported OSes.

The code is already present in the archived repository
(https://github.com/dogtagpki/ldap-sdk-legacy) so it could be removed
from here to make more clear there is not active work on that.