Add Maven build test
A new test has been added to build the code with Maven and Ant
then verify that the artifacts are identical. In the future
the support for Ant will be dropped.

The existing RPM test has been updated to verify that the
artifacts built with XMvn and shipped in RPM are identical to
the ones built directly with Maven.

The build.sh has been updated to provide options to specify
the paths to SLF4J and JSS libraries.